chore(@clayui/pagination): rename ClayPaginationWithBar to PaginationWithBar
